+++ gcpy/__init__.py
@@ -6,22 +6,16 @@ from .benchmark import *
 from .examples import *
 
 from .append_grid_corners import *
-from .benchmark_funcs import *
-from .budget_ox import *
-from .budget_tt import *
 from .constants import *
 from .cstools import *
 from .date_time import *
 from .file_regrid import *
 from .grid import *
 from .grid_stretching_transforms import *
-from .mean_oh_from_logs import *
-from .oh_metrics import *
 from .plot import *
 from .raveller_1D import *
 from .regrid import *
 from .regrid_restart_file import *
-from .ste_flux import *
 from .units import *
 from .util import *
 from ._version import *